digistore/requirements.txt
altgraph==0.17
asgiref==3.2.10
autobahn==20.4.3
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
kiwisolver==1.2.0
markdown2==2.3.9
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.4
pefile==2019.4.18
Pillow==7.2.0
playsound==1.2.2
protobuf==4.0.0rc2
psycopg2==2.8.5
py3dns==3.2.1
pycparser==2.20
PyInstaller==3.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytube==9.6.0
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.24.0
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
txaio==20.4.1
urllib3==1.25.9
validate-email==1.3
whitenoise==5.1.0
django-cloudinary-storage==0.3.0